Fix the cmd/opt swap issue
LGUI should be cmd, and LALT should be opt, but wasn't. Turns out it was some modification stored in eeprom. I saw that this could be the case in the QMK docs and added a button for clearing eeprom. After pressing that button, LGUI is now cmd and LALT opt so I could swap them back.